+++ openrouteservice-api-tests/src/test/java/heigit/ors/services/routing/ResultTest.java
@@ -409,7 +409,7 @@ public class ResultTest extends ServiceTest {
 	}
 
 	@Test
-	public void expectCarToRejectBikeParams() {
+	public void expectCarToRejectProfileParams() {
 
 		// options for cycling profiles
 		JSONObject options = new JSONObject();
@@ -428,7 +428,7 @@ public class ResultTest extends ServiceTest {
 				.get(getEndPointName())
 				.then()
 				.assertThat()
-				.statusCode(200);
+				.statusCode(400);
 	}
 
 	@Test
@@ -787,10 +787,16 @@ http://localhost:8080/ors/routes?
 				.assertThat()
 				.body("any { it.key == 'routes' }", is(true))
 				.body("routes[0].containsKey('extras')", is(true))
-				.body("routes[0].extras.tollways.values.size()", is(1))
+				.body("routes[0].extras.tollways.values.size()", is(3))
 				.body("routes[0].extras.tollways.values[0][0]", is(0))
-				.body("routes[0].extras.tollways.values[0][1]", is(86))
+				.body("routes[0].extras.tollways.values[0][1]", is(52))
 				.body("routes[0].extras.tollways.values[0][2]", is(0))
+                .body("routes[0].extras.tollways.values[1][0]", is(52))
+                .body("routes[0].extras.tollways.values[1][1]", is(66))
+                .body("routes[0].extras.tollways.values[1][2]", is(1))
+                .body("routes[0].extras.tollways.values[2][0]", is(66))
+                .body("routes[0].extras.tollways.values[2][1]", is(86))
+                .body("routes[0].extras.tollways.values[2][2]", is(0))
 				.statusCode(200);
 
 		checkExtraConsistency(response);
@@ -1077,6 +1083,41 @@ http://localhost:8080/ors/routes?
 				.statusCode(200);
 	}
 
+    @Test
+    public void testHGVAxleLoadRestriction() {
+        given()
+                .param("coordinates", "8.686849,49.406093|8.687525,49.405437")
+                .param("instructions", "false")
+                .param("preference", "shortest")
+                .param("profile", "driving-hgv")
+                .param("options", "{\"profile_params\":{\"restrictions\":{\"axleload\":\"12.9\"}},\"vehicle_type\":\"hgv\"}")
+                .param("units", "m")
+                .when().log().ifValidationFails()
+                .get(getEndPointName())
+                .then()
+                .assertThat()
+                .body("any { it.key == 'routes' }", is(true))
+                .body("routes[0].summary.distance", is(132.9f))
+                .body("routes[0].summary.duration", is(44.3f))
+                .statusCode(200);
+
+        given()
+                .param("coordinates", "8.686849,49.406093|8.687525,49.405437")
+                .param("instructions", "true")
+                .param("preference", "shortest")
+                .param("profile", "driving-hgv")
+                .param("options", "{\"profile_params\":{\"restrictions\":{\"axleload\":\"13.1\"}},\"vehicle_type\":\"hgv\"}")
+                .param("units", "m")
+                .when().log().ifValidationFails()
+                .get(getEndPointName())
+                .then()
+                .assertThat()
+                .body("any { it.key == 'routes' }", is(true))
+                .body("routes[0].summary.distance", is(364.3f))
+                .body("routes[0].summary.duration", is(92.7f))
+                .statusCode(200);
+    }
+
 	@Test
 	public void testCarDistanceAndDuration() {
 		// Generic test to ensure that the distance and duration dont get changed
@@ -1258,8 +1299,8 @@ http://localhost:8080/ors/routes?
 				.then()
 				.assertThat()
 				.body("any { it.key == 'routes' }", is(true))
-				.body("routes[0].summary.distance", is(594.4f))
-				.body("routes[0].summary.duration", is(493.8f))
+				.body("routes[0].summary.distance", is(591.6f))
+				.body("routes[0].summary.duration", is(498.7f))
 				.statusCode(200);
 
 		given()
@@ -1384,6 +1425,25 @@ http://localhost:8080/ors/routes?
                 .body("any { it.key == 'routes' }", is(true))
                 .body("routes[0].containsKey('extras')", is(true))
                 .body("routes[0].extras.containsKey('osmId')", is(true))
+                .statusCode(200);
+    }
+
+    @Test
+    public void testAccessRestrictionsWarnings() {
+        given()
+                .param("coordinates", "8.675154,49.407727|8.675863,49.407162")
+                .param("preference", "shortest")
+                .param("profile", getParameter("carProfile"))
+                .when()
+                .get(getEndPointName())
+                .then()
+                .assertThat()
+                .body("any { it.key == 'routes' }", is(true))
+                .body("routes[0].containsKey('warnings')", is(true))
+                .body("routes[0].warnings[0].code", is(1))
+                .body("routes[0].containsKey('extras')", is(true))
+                .body("routes[0].extras.containsKey('roadaccessrestrictions')", is(true))
+                .body("routes[0].extras.roadaccessrestrictions.values[1][2]", is(32))
                 .statusCode(200);
     }
 }